Debugging
Updated:
UI Class  - Adding Phone Number will be storing a long type instead of int

Phonebook class - Phone Number will be storing a long type

PhoneNumber class -  will be storing long type

PersonalInformation class - fixed getEmailAddress method from null error

+++ src/information/Information.java
@@ -5,7 +5,7 @@ public interface Information {
 	
 	public void setPhoneNumber(PhoneNumber number);
 	public void setEmailAddress(String email);
-	public int getPhoneNumber();
+	public long getPhoneNumber();
 	public String getEmailAddress();
